fix(ci): python coverage shard - split oversized executor test shard
